backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Vector Database
qdrant-client>=1.12.0

# OpenAI
openai>=1.55.0

# Utilities
python-slugify>=8.0.4
aiofiles>=24.1.0
